Refa: App and menu delegates
