+++ swama/Sources/SwamaKit/MenuDelegate.swift
@@ -1,9 +1,15 @@
+//
+//  MenuDelegate.swift
+//  swama
+//
+//  Created by haol-co on 14/08/2025
+//
+
 import AppKit
 import Foundation
 
 // MARK: - CLIToolStatus
 
-/// AppDelegate for Swama menu‑bar application
 public enum CLIToolStatus {
     case notInstalled
     case needsUpdate
@@ -23,45 +29,19 @@ private struct CLIToolPaths {
     }
 }
 
-// MARK: - AppDelegate
+// MARK: ‑ MenuDelegate
 
 @MainActor
-public class AppDelegate: NSObject, NSApplicationDelegate {
+public class MenuDelegate: ObservableObject {
     // MARK: Lifecycle
 
     // MARK: ‑ Init
-
-    override public init() {
-        super.init()
-        NSLog("SwamaKit.AppDelegate: init() called.")
-
-        self.serverManager = ServerManager(host: "0.0.0.0", port: 28100)
-    }
+    public init() {}
 
     // MARK: Public
 
-    // MARK: ‑ NSApplicationDelegate
-
-    public func applicationDidFinishLaunching(_: Notification) {
-        // turn app into a menu‑bar‑only application
-        if NSApp.activationPolicy() != .accessory {
-            NSApp.setActivationPolicy(.accessory)
-        }
-
-        // start backend server
-        Task {
-            do { try serverManager?.startInBackground() }
-            catch { NSLog("SwamaKit.AppDelegate: server failed to start → \(error)") }
-        }
-    }
-
-    public func applicationWillTerminate(_: Notification) {
-        Task { await serverManager?.stop() }
-    }
-
     // MARK: Private
 
-    private var serverManager: ServerManager?
 
     // MARK: ‑ CLI Tool Paths
